Find if journals are available in SHERPA/RoMEO

Find if journals are available in SHERPA/RoMEO

rr_journal_find(name, qtype = c("exact", "contains", "starts"),
  key = NULL)

Arguments

name
[character(1+)]
one or several strings to match the titles of the journals

qtype
[character(1)]
in: * "exact" full title must be exactly to provided name, * "contains" the provided name must appear anywhere in the title of the journal, * "starts" the provided name must appear at the start of title of the journal.

key
[character(1)]
a character string containing the API key or NULL (see Details section on how to specify it)

Value

Returns a data frame:

title [character(1)]
the name of the journal

issn [character(1)]
the ISSN of the journal

Details

There are several ways to provide your API key. The best way to know about them is to refer to the vignette about "Setting Up Your API key" accessible with the following command: vignette("setting_up_api_key", package = "rromeo"). You can also use rr_auth that will use the provided key to store it as an environmental variable.

Examples

rr_journal_find(name = "Biostatistics", qtype = "contains")

#> 17 journals match your query terms.

#> Only titles and ISSNs of journals returned. Get more information using `rr_journal_name()`

#>                                                             title      issn
#> 1                               American Journal of Biostatistics 1948-9889
#> 2                          Annals of Biometrics and Biostatistics 2374-0116
#> 4                             Austin Biometrics and Biostatistics 2378-9840
#> 5                Biometrics & biostatistics international journal      <NA>
#> 6                                                   Biostatistics 1465-4644
#> 7                                           Biostatistics -Basel- 1527-2486
#> 8                Biostatistics, bioinformatics and biomathematics 0976-1594
#> 9                                Edorium Journal of Biostatistics      <NA>
#> 10                             Enliven: Biostatistics and Metrics      <NA>
#> 11                   Epidemiology Biostatistics and Public Health 2282-2305
#> 12                         International Journal of Biostatistics 1557-4679
#> 13 International journal of clinical biostatistics and biometrics      <NA>
#> 14                        Journal of Biometrics and Biostatistics 2155-6180
#> 15                      Journal of epidemiology and biostatistics 1359-5229
#> 16                                    JP journal of biostatistics 0973-5143
#> 17                   Monographs in Epidemiology and Biostatistics 0740-0845