Retrieve journal policy using ISSN

Retrieve policy information from the SHERPA/RoMEO API using the ISSN of the journal

rr_journal_issn(issn, key = NULL)

Arguments

issn
character(1+)
one or a vector of journal(s) ISSN(s)

key
character(1)
a character string containing the API key or NULL (see Details section on how to specify it)

Value

Returns a data.frame with the following columns:

title character(1)
the name of the journal

issn character(1)
the ISSN of the journal

romeocolour character(1)
the SHERPA/RoMEO colour of the journal

preprint character(1)
is the preprint (not reviewed) archivable?

postprint character(1)
is the postprint (reviewed but not formatted) archivable?

pdf character(1)
is the publisher's version (reviewed and formatted)

pre_embargo character(1)
if applicable the embargo period before the author(s) can archive the preprint

post_embargo character(1)
if applicable the embargo period before the author(s) can archive the postprint

pdf_embargo character(1)
if applicable the embargo period before the author(s) can archive the publisher's version

Details

To provide your API key to rromeo you can:

provide it as a character string as the key arguments of rromeo functions as rr_*(..., key = "my_key_as_a_string") (NOT RECOMMENDED)

you can define the variable SHERPAROMEO_KEY in an .Renviron file in your working directory, the file should contain the following line SHERPAROMEO_KEY=my_key_without_quotes

you can also define the variable SHERPAROMEO_KEY in an .Rprofile file in your working directory, the file should contain the following line SHERPAROMEO_KEY="my_key_with_quotes"

Examples

rr_journal_issn(issn = "1947-6264")

#>                                                       title      issn
#> 1 A Critical Introduction to Media and Communication Theory 1947-6264
#>   romeocolour preprint  postprint        pdf pre_embargo post_embargo
#> 1      yellow      can restricted restricted        <NA>    12 months
#>   pdf_embargo
#> 1   12 months

rr_journal_issn(issn = c("1947-6264", "0030-1299"))

#>                                                       title      issn
#> 1 A Critical Introduction to Media and Communication Theory 1947-6264
#> 2                                                     Oikos 0030-1299
#>   romeocolour preprint  postprint        pdf pre_embargo post_embargo
#> 1      yellow      can restricted restricted        <NA>    12 months
#> 2      yellow      can restricted     cannot        <NA>    12 months
#>   pdf_embargo
#> 1   12 months
#> 2        <NA>