Retrieve journals policies by matching title

Note that SHERPARoMEO will not return more than 50 journals in a single query. The function will warn you if you are in this case.

rr_journal_name(name, qtype = c("exact", "contains", "starts"),
  key = NULL)

Arguments

name
[character(1+)]
one or several strings to match the titles of the journals

qtype
[character(1)]
in: * "exact" full title must be exactly to provided name, * "contains" the provided name must appear anywhere in the title of the journal, * "starts" the provided name must appear at the start of title of the journal.

key
[character(1)]
a character string containing the API key or NULL (see Details section on how to specify it)

Value

Returns a data.frame with complete information from journal:

title [character(1)]
the name of the journal

issn [character(1)]
the ISSN of the journal

romeocolour [character(1)]
the SHERPA/RoMEO colour of the journal

preprint [character(1)]
is the preprint (not reviewed) archivable?

postprint [character(1)]
is the postprint (reviewed but not formatted)?

pdf [character(1)]
is the publisher's version (reviewed and formatted)

pre_embargo [character(1)]
if applicable the embargo period before the author(s) can archive the preprint

post_embargo [character(1)]
if applicable the embargo period before the author(s) can archive the postprint

pdf_embargo [character(1)]
if applicable the embargo period before the author(s) can archive the publisher's version

Details

There are several ways to provide your API key. The best way to know about them is to refer to the vignette about "Setting Up Your API key" accessible with the following command: vignette("setting_up_api_key", package = "rromeo"). You can also use rr_auth that will use the provided key to store it as an environmental variable.

Examples

rr_journal_name(name = "Journal of Geology")

#>                title      issn romeocolour preprint postprint pdf pre_embargo
#> 1 Journal of Geology 0022-1376       green      can       can can        <NA>
#>   post_embargo pdf_embargo
#> 1         <NA>        <NA>

rr_journal_name(name = "Biogeography", qtype = "contains")

#> 5 journals match your query terms.

#> Recursively fetching data from each journal. This may take some time...

#>                                     title      issn romeocolour preprint
#> 1                            Biogeography 1345-0662        <NA>     <NA>
#> 2               Frontiers of Biogeography 1948-6596        <NA>     <NA>
#> 3         Global Ecology and Biogeography 1466-822X      yellow      can
#> 4 Global Ecology and Biogeography Letters 0960-7447        <NA>     <NA>
#> 5                 Journal of Biogeography 0305-0270      yellow      can
#>    postprint    pdf pre_embargo post_embargo pdf_embargo
#> 1       <NA>   <NA>        <NA>         <NA>        <NA>
#> 2       <NA>   <NA>        <NA>         <NA>        <NA>
#> 3 restricted cannot        <NA>    12 months        <NA>
#> 4       <NA>   <NA>        <NA>         <NA>        <NA>
#> 5 restricted cannot        <NA>    12 months        <NA>

# You can also query multiple journals with exact titles in a single call
rr_journal_name(name = c("Journal of Biogeography", "PLoS ONE"),
                qtype = "exact")

#>                     title      issn romeocolour preprint  postprint    pdf
#> 1 Journal of Biogeography 0305-0270      yellow      can restricted cannot
#> 2                PLoS ONE 1932-6203       green      can        can    can
#>   pre_embargo post_embargo pdf_embargo
#> 1        <NA>    12 months        <NA>
#> 2        <NA>         <NA>        <NA>