Query publisher by RoMEO colour

SHERPA/RoMEO classifies publisher in different colours depending on their archiving policies.

green publishers let authors archive preprint and postprint or publisher's version/PDF,

blue publishers let authors archive postprint or publisher's version/PDF,

yellow publishers let authors archive preprint,

white publishers do not formally support archival.

rr_romeo_colour(romeo_colour = c("green", "blue", "yellow", "white"),
  key = NULL)

Arguments

romeo_colour
[character(1)]
in c("green", "blue", "yellow", "white") the SHERPA/RoMEO colour to retrieve

key
[character(1)]
a character string containing the API key or NULL (see Details section on how to specify it)

Value

Returns a data frame with the following columns:

romeoid [integer(1)]
the internal index of the publisher in the SHERPA/RoMEO database

publisher [character(1)]
the name of the publisher

alias [character(1)]
if applicable an alternative name of the publisher or the name of the specific publishing branch

romeocolour [character(1)]
a colour assigned by the database that reflects the default policies of the publisher

preprint [character(1)]
is the preprint (not reviewed) archivable?

postprint [character(1)]
is the postprint (reviewed but not formatted) archivable?

pdf [character(1)]
is the publisher's version (reviewed and formatted) archivable?

Details

For more details about the definitions of RoMEO colours check the FAQ section of SHERPA/RoMEO

Note that when using rr_romeo_colour() the API returns all the publishers in the selected category, so the results are generally bigger in size than specific functions like [rr_journal_name()] or [rr_publisher_id()]

Examples

# NOT RUN {
rr_romeo_colour(romeo_colour = "green")
# }