Get Publisher Policy by Publisher's Continent

Retrieve publisher's policy based on publisher's continent. This function does not work for unclassified or international publishers.

rr_publisher_continent(continent = c("Africa", "Antarctica", "Asia",
  "Australasia", "Caribbean", "Central America", "Europe", "North America",
  "Oceania", "South America"), key = NULL)

Arguments

continent
character(1+)
one or a vector of strings in c("Africa", "Antarctica", "Asia", "Australasia", "Carribean", "Central America", "Europe", "North America", "Oceania", "South America")
the continent name to retrieve

key
character(1)
a character string containing the API key or NULL (see Details section on how to specify it)

Value

Returns a data frame with the following columns:

romeoid integer(1)
the internal index of the publisher in the SHERPA/RoMEO database

publisher character(1)
the name of the publisher

alias character(1)
if applicable an alternative name of the publisher or the name of the specific publishing branch

romeocolour character(1)
a colour assigned by the database that reflects the default policies of the publisher

preprint character(1)
is the preprint (not reviewed) archivable?

postprint character(1)
is the postprint (reviewed but not formatted) archivable?

pdf character(1)
is the publisher's version (reviewed and formatted) archivable?

Details

To provide your API key to rromeo you can:

provide it as a character string as the key arguments of rromeo functions as rr_*(..., key = "my_key_as_a_string") (NOT RECOMMENDED)

you can define the variable SHERPAROMEO_KEY in an .Renviron file in your working directory, the file should contain the following line SHERPAROMEO_KEY=my_key_without_quotes

you can also define the variable SHERPAROMEO_KEY in an .Rprofile file in your working directory, the file should contain the following line SHERPAROMEO_KEY="my_key_with_quotes"

Examples

rr_publisher_continent(continent = "Caribbean")

#>    romeoid
#> 1     2192
#> 2     2687
#> 3     2614
#> 4     1371
#> 5     1717
#> 6     2942
#> 7     2815
#> 8     2034
#> 9     1280
#> 10    1505
#> 11    2620
#>                                                                                                                     publisher
#> 1                                                                             Instituto de Investigaciones de Sanidad Vegetal
#> 2                               Instituto Superior Politécnico José Antonio Echeverría-Cujae, Facultad de Ingeniería Mecánica
#> 3                                        Sistema Universitario Ana G. Mendez. Universidad Metropolitana, Escuela de Educación
#> 4                                                                                                   Sociedad Cubana de Física
#> 5                                                                                Sociedad Cubana de Neurología y Neurocirugía
#> 6  Universidad Central Marta Abreu de Las Villas, Facultad de Ciencias Agropecuarias, Centro de Investigaciones Agropecuarias
#> 7                                                                              Universidad de La Habana, Facultad de Biologia
#> 8                                                                                    Universidad de las Ciencias Informaticas
#> 9                                                                   Universidad de Las Tunas, Centro de Estudios de Didáctica
#> 10                                                                                                     Universidad de Oriente
#> 11                                                                    Universidad de Puerto Rico, Recinto de Ciencias Médicas
#>    alias romeocolour preprint postprint  pdf
#> 1   <NA>       green      can       can  can
#> 2   <NA>        blue   cannot       can  can
#> 3   <NA>       green      can    cannot  can
#> 4   <NA>       green      can       can <NA>
#> 5   <NA>        blue   cannot    cannot  can
#> 6   <NA>        blue   cannot    cannot  can
#> 7   <NA>       green      can       can  can
#> 8   <NA>        blue   cannot    cannot  can
#> 9   <NA>       green      can       can  can
#> 10  <NA>       green      can       can <NA>
#> 11  <NA>       green      can       can  can

rr_publisher_continent(continent = "Central America")

#>   romeoid                                                      publisher alias
#> 1     798                         Asociacion Nicaraguense de Entomologia  <NA>
#> 2    2243                Colegio Profesional de Psicologos de Costa Rica  <NA>
#> 3    1364                                      Universidad de Costa Rica  <NA>
#> 4     939 Universidad de Costa Rica, Centro Centroamericano de Población  <NA>
#> 5    1728                             Universidad Nacional de Costa Rica  <NA>
#>   romeocolour preprint postprint  pdf
#> 1        blue   cannot       can  can
#> 2        blue   cannot       can  can
#> 3       green      can       can  can
#> 4        blue   cannot       can <NA>
#> 5       green      can       can  can

rr_publisher_continent(continent = c("Caribbean", "Central America"))

#>    romeoid
#> 1     2192
#> 2     2687
#> 3     2614
#> 4     1371
#> 5     1717
#> 6     2942
#> 7     2815
#> 8     2034
#> 9     1280
#> 10    1505
#> 11    2620
#> 12     798
#> 13    2243
#> 14    1364
#> 15     939
#> 16    1728
#>                                                                                                                     publisher
#> 1                                                                             Instituto de Investigaciones de Sanidad Vegetal
#> 2                               Instituto Superior Politécnico José Antonio Echeverría-Cujae, Facultad de Ingeniería Mecánica
#> 3                                        Sistema Universitario Ana G. Mendez. Universidad Metropolitana, Escuela de Educación
#> 4                                                                                                   Sociedad Cubana de Física
#> 5                                                                                Sociedad Cubana de Neurología y Neurocirugía
#> 6  Universidad Central Marta Abreu de Las Villas, Facultad de Ciencias Agropecuarias, Centro de Investigaciones Agropecuarias
#> 7                                                                              Universidad de La Habana, Facultad de Biologia
#> 8                                                                                    Universidad de las Ciencias Informaticas
#> 9                                                                   Universidad de Las Tunas, Centro de Estudios de Didáctica
#> 10                                                                                                     Universidad de Oriente
#> 11                                                                    Universidad de Puerto Rico, Recinto de Ciencias Médicas
#> 12                                                                                     Asociacion Nicaraguense de Entomologia
#> 13                                                                            Colegio Profesional de Psicologos de Costa Rica
#> 14                                                                                                  Universidad de Costa Rica
#> 15                                                             Universidad de Costa Rica, Centro Centroamericano de Población
#> 16                                                                                         Universidad Nacional de Costa Rica
#>    alias romeocolour preprint postprint  pdf
#> 1   <NA>       green      can       can  can
#> 2   <NA>        blue   cannot       can  can
#> 3   <NA>       green      can    cannot  can
#> 4   <NA>       green      can       can <NA>
#> 5   <NA>        blue   cannot    cannot  can
#> 6   <NA>        blue   cannot    cannot  can
#> 7   <NA>       green      can       can  can
#> 8   <NA>        blue   cannot    cannot  can
#> 9   <NA>       green      can       can  can
#> 10  <NA>       green      can       can <NA>
#> 11  <NA>       green      can       can  can
#> 12  <NA>        blue   cannot       can  can
#> 13  <NA>        blue   cannot       can  can
#> 14  <NA>       green      can       can  can
#> 15  <NA>        blue   cannot       can <NA>
#> 16  <NA>       green      can       can  can