Get Publisher Policy by Publisher Name

Use SHERPA/RoMEO API to retrieve a specific publisher policies on manuscript archival based on matching the name of the publishers.

rr_publisher_name(name, qtype = c("all", "any", "exact"), key = NULL)

Arguments

name
character(1+)
One or a vector of query string(s) to search publisher name

qtype
character(1)
in c("all", "any", "exact") define the type of matching:

all means that all strings in name must appear in any order or location

any means that at least one of the strings in name must appear

exact means that the name string must appear in the publisher's name or its alias.

key
character(1)
a character string containing the API key or NULL (see Details section on how to specify it)

Value

Returns a data frame with the following columns:

romeoid integer(1)
the internal index of the publisher in the SHERPA/RoMEO database

publisher character(1)
the name of the publisher

alias character(1)
if applicable an alternative name of the publisher or the name of the specific publishing branch

romeocolour character(1)
a colour assigned by the database that reflects the default policies of the publisher

preprint character(1)
is the preprint (not reviewed) archivable?

postprint character(1)
is the postprint (reviewed but not formatted) archivable?

pdf character(1)
is the publisher's version (reviewed and formatted) archivable?

Details

To provide your API key to rromeo you can:

provide it as a character string as the key arguments of rromeo functions as rr_*(..., key = "my_key_as_a_string") (NOT RECOMMENDED)

you can define the variable SHERPAROMEO_KEY in an .Renviron file in your working directory, the file should contain the following line SHERPAROMEO_KEY=my_key_without_quotes

you can also define the variable SHERPAROMEO_KEY in an .Rprofile file in your working directory, the file should contain the following line SHERPAROMEO_KEY="my_key_with_quotes"

Examples

rr_publisher_name(name = "Optical Society", qtype = "all")

#>   romeoid                                          publisher alias romeocolour
#> 1     219                           European Optical Society  <NA>        blue
#> 2     108                         Optical Society of America  <NA>       green
#> 3     127 Society of Photo-optical Instrumentation Engineers  SPIE       green
#>   preprint postprint    pdf
#> 1  unclear       can   <NA>
#> 2      can       can cannot
#> 3      can       can    can

rr_publisher_name(name = "Swiss Chemistry", qtype = "any")

#>    romeoid
#> 1     2677
#> 2      318
#> 3      113
#> 4      195
#> 5     2049
#> 6       84
#> 7       43
#> 8      452
#> 9     2118
#> 10      63
#> 11    1649
#> 12    2612
#> 13     411
#> 14     440
#> 15    3020
#> 16     673
#>                                                                                                 publisher
#> 1                                                    Academia de Stiinte a Moldovei, Institutul de Chimie
#> 2                                                             American Association for Clinical Chemistry
#> 3                                                 American Society for Biochemistry and Molecular Biology
#> 4                                                                                       Chemistry Central
#> 5                                                                         EMH Schweizerischer Arzteverlag
#> 6                                                       International Union of Pure and Applied Chemistry
#> 7                                                                           Internet Journal of Chemistry
#> 8                                          Japan Society for Bioscience, Biotechnology, and Agrochemistry
#> 9                                                                                  Nihon Bunseki Kagakkai
#> 10                                                                             Royal Society of Chemistry
#> 11  Russian Academy of Sciences, Siberian Branch, Siberian Institute of Plant Physiology and Biochemistry
#> 12                                                                                                   SBBq
#> 13                                                                  Schweizerische Chemische Gesellschaft
#> 14                                                Society of Environmental Toxicology and Chemistry Press
#> 15                                                                               Swiss Medical Press GmbH
#> 16 Universidade Federal de Mato Grosso do Sul, Department of Chemistry, Center for Science and Technology
#>                                                       alias romeocolour
#> 1    Academy of Sciences of Moldova, Institute of Chemistry       green
#> 2                                                      <NA>      yellow
#> 3                                                      <NA>       green
#> 4                                                      <NA>       green
#> 5                              EMH Swiss Medical Publishers       green
#> 6                                                      <NA>        blue
#> 7                                                      <NA>       green
#> 8  &#26085;&#26412;&#36786;&#33464;&#21270;&#23398;&#20250;       white
#> 9                    Japan Society for Analytical Chemistry        blue
#> 10                                                     <NA>      yellow
#> 11                                                     <NA>       green
#> 12 Brazilian Society for Biochemistry and Molecular Biology       green
#> 13                                  Swiss Chemistry Society       white
#> 14                                                    SETAC       white
#> 15                                                     <NA>       green
#> 16                                                     <NA>       green
#>    preprint  postprint        pdf
#> 1       can        can        can
#> 2       can restricted     cannot
#> 3       can        can restricted
#> 4       can        can        can
#> 5       can        can        can
#> 6    cannot     cannot        can
#> 7       can        can     cannot
#> 8   unclear restricted       <NA>
#> 9   unclear     cannot        can
#> 10      can restricted     cannot
#> 11      can        can        can
#> 12      can        can        can
#> 13  unclear     cannot restricted
#> 14   cannot     cannot       <NA>
#> 15      can        can        can
#> 16      can        can       <NA>

rr_publisher_name(name = "Swiss Chemistry", qtype = "exact")

#>   romeoid                             publisher                   alias
#> 1     411 Schweizerische Chemische Gesellschaft Swiss Chemistry Society
#>   romeocolour preprint postprint        pdf
#> 1       white  unclear    cannot restricted